Move beam_centre device to a more common location and instantiate on i19-1
