+++ src/dodal/beamlines/i04.py
@@ -12,7 +12,6 @@ from dodal.devices.attenuator.attenuator import BinaryFilterAttenuator
 from dodal.devices.backlight import Backlight
 from dodal.devices.baton import Baton
 from dodal.devices.beamlines.i03.dcm import DCM
-from dodal.devices.beamlines.i04.beam_centre import CentreEllipseMethod
 from dodal.devices.beamlines.i04.beamsize import Beamsize
 from dodal.devices.beamlines.i04.constants import RedisConstants
 from dodal.devices.beamlines.i04.max_pixel import MaxPixel
@@ -26,6 +25,7 @@ from dodal.devices.flux import Flux
 from dodal.devices.ipin import IPin
 from dodal.devices.motors import XYZStage
 from dodal.devices.mx_phase1.beamstop import Beamstop
+from dodal.devices.oav.beam_centre.beam_centre import CentreEllipseMethod
 from dodal.devices.oav.oav_detector import (
     OAVBeamCentrePV,
     ZoomControllerWithBeamCentres,
